+++ src/app/components/perfil/perfil.component.html
@@ -1,9 +1,2 @@
-<div *ngIf="userdata">
-    <h1>Bienvenido, {{ userdata.username }}</h1>
-    <p>Email: {{ userdata.email }}</p>
-    <p>Teléfono: {{ userdata.tel }}</p>
-  </div>
-  <div *ngIf="!userdata">
-    <p>No se encontraron datos del cliente.</p>
-  </div>
+<div>perfil</div>
